Move Fix Splits to Settings → Advanced
