+++ packages/desktop-client/src/components/settings/index.js
@@ -16,6 +16,7 @@ import useServerVersion from '../../hooks/useServerVersion';
 import { Page } from '../Page';
 import EncryptionSettings from './Encryption';
 import ExportBudget from './Export';
+import FixSplitsTool from './FixSplits';
 import FormatSettings from './Format';
 import GlobalSettings from './Global';
 import { ResetCache, ResetSync } from './Reset';
@@ -101,6 +102,7 @@ function Settings({
           <AdvancedAbout prefs={prefs} />
           <ResetCache />
           <ResetSync resetSync={resetSync} />
+          <FixSplitsTool />
         </AdvancedToggle>
       </View>
     </Page>
